// server/src/types/Config/Config.ts
import { AutoGeneratedConfig } from './AutoGeneratedConfig';

/**
 * Shape of exported config that will be used through the server.
 *
 * For more info, see the schema in `.github/config.schema.json`.
 *
 * For default values, see `src/constants/defaultConfig.ts`.
 *
 * Note that this is **not** the same as the config in `config.json`,
 * as the latter has values that cannot be represented in JSON form
 * like sets and regular expressions.
 *
 * For the schema of `config.json`, see {@link ImportedConfig}.
 */
export interface Config extends AutoGeneratedConfig {
    port: number;

    clientUrls: Set<string>;

    numProxies: number;

    maxRequestsPerMinute: number;

    rateLimitBypassTokens: Set<string>;

    /** Required value, process will exit if omitted! */
    mongoURI: string;

    /** Process will exit if this is > 38 characters long! */
    mongoDbName: string;

    jwtSecret: string;

    /** Required value, process will exit if omitted! */
    discordClientId: string;

    /** Required value, process will exit if omitted! */
    discordClientSecret: string;
}